+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Media (like images, videos, audios) can be attached to a feed item
+- Author can be attached to a feed or to a feed item
 
 ## [0.1.0] - 2022-04-01